% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_palindrome.R
\name{Make_palindrome}
\alias{Make_palindrome}
\title{Title Converts}
\usage{
Make_palindrome(x)
}
\arguments{
\item{x}{number to be converted to a palindrome by reversing and adding}
}
\value{
all the steps in creating turning a number into a palindrome
}
\description{
Title Converts
}
\examples{
 Make_palindrome(12)
}
